requirements.txt
cryptography==3.1.1
aiohttp==3.6.2
aiohttp-session==2.9.0
PyJWT==2.1.0
aiohttp-jinja2==1.2.0
aiopg==1.2.1
peewee-async==0.7.1
peewee==3.14.4
APScheduler==3.6.3
pytest-aiohttp==0.3.0
